+++ src/Utility/BackupManager.php
@@ -64,9 +64,15 @@ class BackupManager
      * @param int $keep Number of backups that you want to keep
      * @return array<array{filename: string, basename: string, path: string, compression: \DatabaseBackup\Compression, size: int|false, datetime: \Cake\I18n\DateTime}>
      * @throws \InvalidArgumentException With an Invalid rotate value.
+     * @deprecated `BackupManager::rotate()` has been deprecated and will be removed in a future release
      */
     public static function rotate(int $keep): array
     {
+        deprecationWarning(
+            '2.15.0',
+            '`BackupManager::rotate()` has been deprecated and will be removed in a future release'
+        );
+
         if ($keep < 1) {
             throw new InvalidArgumentException(__d('database_backup', 'Invalid `$keep` value'));
         }