+++ tests/TestCase/Utility/BackupManagerTest.php
@@ -23,6 +23,7 @@ use DatabaseBackup\Utility\BackupManager;
 use InvalidArgumentException;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\Attributes\WithoutErrorHandler;
 
 /**
  * BackupManagerTest class.
@@ -52,9 +53,10 @@ class BackupManagerTest extends TestCase
     }
 
     #[Test]
+    #[WithoutErrorHandler]
     public function testRotate(): void
     {
-        $this->assertSame([], BackupManager::rotate(1));
+        $this->deprecated(fn () => $this->assertSame([], BackupManager::rotate(1)));
 
         /**
          * Creates 3 backups (`$initialFiles`) and keeps only 2 of them.
@@ -63,16 +65,19 @@ class BackupManagerTest extends TestCase
          */
         $initialFiles = $this->createSomeBackups();
 
-        $rotate = BackupManager::rotate(2);
-        $this->assertCount(1, $rotate);
-        $this->assertSame($initialFiles[2], $rotate[0]['path']);
+        $this->deprecated(function () use ($initialFiles) {
+            $rotate = BackupManager::rotate(2);
+            $this->assertCount(1, $rotate);
+            $this->assertSame($initialFiles[2], $rotate[0]['path']);
+        });
     }
 
     #[Test]
+    #[WithoutErrorHandler]
     public function testRotateWithInvalidKeepValue(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Invalid `$keep` value');
-        BackupManager::rotate(-1);
+        $this->deprecated(fn () => BackupManager::rotate(-1));
     }
 }